Increasing the throughput of getting valueProviders by allowing concurrent read/writes on the map without locking the entire collection.
 
# After:
Warmup Iteration   1: 24720.953 ±(99.9%) 1990.813 ms/op
Warmup Iteration   2: 24258.182 ±(99.9%) 1285.860 ms/op
Iteration   1: 25135.823 ±(99.9%) 1196.562 ms/op
Iteration   2: 23966.300 ±(99.9%) 1276.853 ms/op
Iteration   3: 25538.036 ±(99.9%) 1964.715 ms/op
Iteration   4: 26969.966 ±(99.9%) 2155.508 ms/op
Iteration   5: 26302.678 ±(99.9%) 3366.828 ms/op

Result "com.sabre.oss.yare.performance.suits.MultithreadedTest.benchmark":
  25582.560 ±(99.9%) 4414.541 ms/op [Average]
  (min, avg, max) = (23966.300, 25582.560, 26969.966), stdev = 1146.442
  CI (99.9%): [21168.019, 29997.102] (assumes normal distribution)


# Before:  
Warmup Iteration   1: 30711.778 ±(99.9%) 3640.221 ms/op
Warmup Iteration   2: 34629.744 ±(99.9%) 4227.998 ms/op
Iteration   1: 30236.826 ±(99.9%) 2876.878 ms/op
Iteration   2: 30757.804 ±(99.9%) 3400.727 ms/op
Iteration   3: 30518.556 ±(99.9%) 3088.247 ms/op
Iteration   4: 33057.338 ±(99.9%) 2871.403 ms/op
Iteration   5: 30440.224 ±(99.9%) 3197.480 ms/op

Result "com.sabre.oss.yare.performance.suits.MultithreadedTest.benchmark":
  31002.149 ±(99.9%) 4481.845 ms/op [Average]
  (min, avg, max) = (30236.826, 31002.149, 33057.338), stdev = 1163.921
  CI (99.9%): [26520.305, 35483.994] (assumes normal distribution)

![lock](https://user-images.githubusercontent.com/12997861/63272148-d1f93380-c29b-11e9-81e1-fb146dafab47.PNG)