@CLAassistant 2019-08-19
[![CLA assistant check](https://cla-assistant.io/pull/badge/not_signed)](https://cla-assistant.io/SabreOSS/yare?pullRequest=69) <br/>Thank you for your submission, we really appreciate it. Like many open source projects, we ask that you all sign our [Contributor License Agreement](https://cla-assistant.io/SabreOSS/yare?pullRequest=69) before we can accept your contribution.<br/>**0** out of **2** committers have signed the CLA.<br/><br/>:x: Wiktor Klosowski<br/>:x: WiktorKlosowski<br/><hr/>**Wiktor Klosowski** seems not to be a GitHub user. You need a GitHub account to be able to sign the CLA. If you have already a GitHub account, please [add the email address used for this commit to your account](https://help.github.com/articles/why-are-my-commits-linked-to-the-wrong-user/#commits-are-not-linked-to-any-user).<br/><sub>You have signed the CLA already but the status is still pending? Let us [recheck](https://cla-assistant.io/check/SabreOSS/yare?pullRequest=69) it.</sub>