@MarcinLachowicz 2019-08-20
Hello,
thank you for contribution. Please fix build: https://travis-ci.org/SabreOSS/yare/builds/574219089