@krzysztofreczek 2019-08-20
@WiktorKlosowski, well done. I am wondering would it make sense to check how does the number of facts and rules impact the performance? I guess you can simply parameterize (with @Param) the benchmark you have created.