@WiktorKlosowski 2019-08-20
@krzysztofreczek Good idea with parameters. 
Performance is mostly affected by number of "." used in rules.

# After
Benchmark                    (numberOfFacts)  (numberOfRules)  (ruleComplexity)  Mode  Cnt     Score     Error  Units
MultithreadedTest.benchmark               50               50                 5  avgt    5    13.149 ±   2.904  ms/op
MultithreadedTest.benchmark               50               50                50  avgt    5   102.621 ±   4.640  ms/op
MultithreadedTest.benchmark               50               50               500  avgt    5   998.894 ± 134.735  ms/op
MultithreadedTest.benchmark               50              100                 5  avgt    5    30.724 ±   4.730  ms/op
MultithreadedTest.benchmark               50              100                50  avgt    5   234.987 ±   9.497  ms/op
MultithreadedTest.benchmark               50              100               500  avgt    5  2155.567 ± 101.030  ms/op
MultithreadedTest.benchmark              100               50                 5  avgt    5    19.995 ±   2.427  ms/op
MultithreadedTest.benchmark              100               50                50  avgt    5   131.471 ±  17.715  ms/op
MultithreadedTest.benchmark              100               50               500  avgt    5  1041.247 ±  21.014  ms/op
MultithreadedTest.benchmark              100              100                 5  avgt    5    47.441 ±   5.982  ms/op
MultithreadedTest.benchmark              100              100                50  avgt    5   330.313 ±   4.994  ms/op
MultithreadedTest.benchmark              100              100               500  avgt    5  2330.488 ±  40.036  ms/op

# Before
Benchmark                    (numberOfFacts)  (numberOfRules)  (ruleComplexity)  Mode  Cnt     Score     Error  Units
MultithreadedTest.benchmark               50               50                 5  avgt    5    19.579 ±   2.679  ms/op
MultithreadedTest.benchmark               50               50                50  avgt    5   167.538 ±   8.705  ms/op
MultithreadedTest.benchmark               50               50               500  avgt    5  1658.271 ± 113.553  ms/op
MultithreadedTest.benchmark               50              100                 5  avgt    5    41.458 ±   0.692  ms/op
MultithreadedTest.benchmark               50              100                50  avgt    5   358.604 ±  11.421  ms/op
MultithreadedTest.benchmark               50              100               500  avgt    5  3512.990 ± 368.452  ms/op
MultithreadedTest.benchmark              100               50                 5  avgt    5    22.135 ±   2.213  ms/op
MultithreadedTest.benchmark              100               50                50  avgt    5   181.495 ±   9.768  ms/op
MultithreadedTest.benchmark              100               50               500  avgt    5  1714.727 ±  88.329  ms/op
MultithreadedTest.benchmark              100              100                 5  avgt    5    51.023 ±   3.031  ms/op
MultithreadedTest.benchmark              100              100                50  avgt    5   478.115 ±  67.561  ms/op
MultithreadedTest.benchmark              100              100               500  avgt    5  3650.446 ± 815.868  ms/op